Scalar mutations are working
