+++ geode-graphql/src/test/java/org/g2ql/graphql/GraphqlExecutorIntegrationTest.java
@@ -110,11 +110,12 @@ public class GraphqlExecutorIntegrationTest {
 
   @Test
   public void testGraphQLPutFooKeyValue() throws IOException {
-    String m = "{\"query\":\"mutation PutFoo($key: String, $value: String) {\\n  putFoo(key: $key, value: $value)\\n}\\n\",\"variables\":{\"key\":\"3\",\"value\":\"Three\"},\"operationName\":\"PutFoo\"}";
+    String m =
+        "{\"query\":\"mutation PutFoo($key: String, $value: String) {\\n  putFoo(key: $key, value: $value)\\n}\\n\",\"variables\":{\"key\":\"3\",\"value\":\"Three\"},\"operationName\":\"PutFoo\"}";
     HttpResponse response = Request.Post("http://localhost:3000/graphql")
         .bodyString(m, ContentType.TEXT_PLAIN).execute().returnResponse();
     String responseString = new BasicResponseHandler().handleResponse(response);
-    assertThat(responseString).isEqualTo("{\"data\":{\"Foo\":\"One\"}}");
+    assertThat(responseString).isEqualTo("{\"data\":{\"putFoo\":\"Three\"}}");
   }
 
   @Test
